backend/requirements.txt
# Core dependencies with exact working versions
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.3
matplotlib==3.10.3
joblib==1.5.0

# RDKit and chemistry dependencies
rdkit==2025.3.2
rdkit-pypi==2022.9.5

# Deep Learning dependencies
torch==2.7.0
torch-geometric==2.6.1
numba==0.61.2
llvmlite==0.44.0

# Web framework dependencies
flask==3.1.1
flask-cors==5.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0
click==8.2.0

# Data processing and utilities
requests==2.32.3
python-dotenv==1.0.0
pydantic==1.10.12
tqdm==4.67.1
packaging==25.0

# Jupyter/IPython dependencies (if needed for development)
ipython==8.36.0
ipykernel==6.29.5
jupyter_client==8.6.3
jupyter_core==5.7.2

# Additional scientific dependencies
seaborn==0.13.2
spectrum-utils==0.4.2
pyteomics==4.7.5
networkx==3.4.2
sympy==1.14.0

# Support libraries
pillow==11.2.1
fonttools==4.58.0
contourpy==1.3.2
cycler==0.12.1
kiwisolver==1.4.8
matplotlib-inline==0.1.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2

# Development and debugging tools
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
jedi==0.19.2
parso==0.8.4
pure_eval==0.2.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13

# System and platform dependencies
appnope==0.1.4
platformdirs==4.3.8
psutil==7.0.0
threadpoolctl==3.6.0
typing_extensions==4.13.2

# Terminal and prompt dependencies
pexpect==4.9.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
Pygments==2.19.1

# Async and networking
tornado==6.4.2
pyzmq==26.4.0
nest-asyncio==1.6.0

# Communication and misc
comm==0.2.2
asttokens==3.0.0
exceptiongroup==1.3.0


# Production WSGI Server
gunicorn==21.2.0

# Additional dependencies for other features
fastobo==0.13.0
lark==1.2.2
dill==0.4.0
treelib==1.7.1

# Azure and cloud dependencies (if needed)
azure-core==1.34.0
azure-storage-blob==12.25.1
fsspec==2025.3.2

# HTTP and web dependencies
aiohttp==3.11.18
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
urllib3==2.4.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
yarl==1.20.0
frozenlist==1.6.0

# Cryptography and security
cryptography==45.0.2
cffi==1.17.1
pycparser==2.22

# HTTP clients
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.9.0
sniffio==1.3.1

# Validation and typing
annotated-types==0.7.0
pydantic_core==2.33.2
typing-inspection==0.4.0

# File and system utilities
appdirs==1.4.4
filelock==3.18.0
distro==1.9.0
isodate==0.7.2
